chore: refactor hooks to CRUD and remove legacy technical debt
- Architecture: Renamed 'hooks/new' to 'hooks/CRUD' to standardize V3 data layer.
- Cleanup: Deleted deprecated UI components (oldAnalytics, oldGoals, oldExport).
- Tech Debt: Removed redundant hook definitions and the 'redundent' folder.
- Config: Updated 'next.config.ts' and Service Worker to reflect new structure.
- Validated build and auto-updated imports.

Closes #18
Ref #37
