--- hooks/useCreateSession.ts
@@ -1,99 +1,0 @@
-// hooks/useCreateSession.ts
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { addDoc, collection, Timestamp } from 'firebase/firestore';
-import { db, auth } from '@/lib/firebase';
-import { Session } from '@/types'; // Our v2 UI type
-import { nanoid } from 'nanoid'; // Let's use nanoid like v0
-import { toast } from 'sonner';
-
-// This is the data type we'll save to Firestore
-
-// --- 1. THIS IS THE FIX ---
-// This interface now matches the 'finalV0Data' object
-// that SessionTracker.tsx will create.
-interface FinalV0DataInput {
-  created_at?: Timestamp | '';
-  id: string;
-  userId: string;
-  title: string;
-  session_type_id: string;
-  notes: string;
-  breaks: any[]; // Or be more specific: Break[]
-  started_at: string;
-  ended_at: string;
-  total_focus_ms: number;
-  total_break_ms: number;
-}
-// --- Query Client Helper ---
-const useSessionQueryClient = () => {
-  const queryClient = useQueryClient();
-  // We only invalidate ['sessions']
-  return () => queryClient.invalidateQueries({ queryKey: ['sessions'] });
-};
-
-// --- Create ---
-// The async function that does the server work
-const createSessionOnFirebase = async (sessionData: Partial<FinalV0DataInput>) => {
-  const user = auth.currentUser;
-  if (!user) throw new Error('No authenticated user found');
-
-  const dataToSave = {
-    id: user.uid,
-    userId: user.uid,
-    title: sessionData.title,
-    session_type_id: sessionData.session_type_id,
-    notes: sessionData.notes,
-    breaks: sessionData.breaks,
-    started_at: sessionData.started_at,
-    ended_at: sessionData.ended_at,
-    total_focus_ms: sessionData.total_focus_ms,
-    total_break_ms: sessionData.total_break_ms,
-    // created_at: Timestamp || '',
-  };
-
-  const docRef = await addDoc(collection(db, 'sessions'), dataToSave);
-  return docRef.id;
-};
-
-// The custom hook that our component will use
-export const useCreateSession = () => {
-  const invalidateSessions = useSessionQueryClient();
-
-  return useMutation({
-    mutationFn: createSessionOnFirebase,
-
-    onSuccess: () => {
-      toast.success('Session saved');
-      console.log('Session saved to Firebase "sessions" collection.');
-      // invalidateSessions();
-    },
-    onError: (error) => {
-      toast.error('Failed to save session', { description: error.message });
-      console.error('Failed to save session (will retry):', error);
-      // We don't reset the store here, so the user can try again
-    },
-  });
-};
-
-
-/*
-* --- Example Usage On Call Site --- 
-
-const createSessionMutation = useCreateSession(); // Instantiate the mutations
-
-const finalV0Data = {
-            // id: user?.uid, // UUID will be generated by Firestore
-            id: nanoid(),
-            userId: user.uid,
-            title: title,
-            session_type_id: type,
-            notes: notes || "",
-            breaks: breaks,
-            started_at: sessionStartTime,
-            ended_at: endTime,
-            total_focus_ms: sessionTime,
-            total_break_ms: breakTime,
-        };
-        
-createSessionMutation.mutateAsync(finalV0Data);
-*/
