+++ public/sw.js
@@ -1,1 +1,1 @@
-if(!self.define){let e,s={};const a=(a,c)=>(a=new URL(a+".js",c).href,s[a]||new Promise(s=>{if("document"in self){const e=document.createElement("script");e.src=a,e.onload=s,document.head.appendChild(e)}else e=a,importScripts(a),s()}).then(()=>{let e=s[a];if(!e)throw new Error(`Module ${a} didn’t register its module`);return e}));self.define=(c,n)=>{const t=e||("document"in self?document.currentScript.src:"")||location.href;if(s[t])return;let i={};const r=e=>a(e,t),d={module:{uri:t},exports:i,require:r};s[t]=Promise.all(c.map(e=>d[e]||r(e))).then(e=>(n(...e),i))}}define(["./workbox-3c9d0171"],function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/7aBLBtj6MpjVyKPlH88ai/_buildManifest.js",revision:"69876e280c6cdb936db78c7688725a56"},{url:"/_next/static/7aBLBtj6MpjVyKPlH88ai/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/14-11e4eba3370dba8c.js",revision:"11e4eba3370dba8c"},{url:"/_next/static/chunks/161cdf59-bdcc1c65f6aee355.js",revision:"bdcc1c65f6aee355"},{url:"/_next/static/chunks/217-06d976e7d546d610.js",revision:"06d976e7d546d610"},{url:"/_next/static/chunks/282-a23421dd9b42535e.js",revision:"a23421dd9b42535e"},{url:"/_next/static/chunks/341-34f6c6c7b1211a4c.js",revision:"34f6c6c7b1211a4c"},{url:"/_next/static/chunks/351-3bcffb9f24019b3b.js",revision:"3bcffb9f24019b3b"},{url:"/_next/static/chunks/371-955b35777a894855.js",revision:"955b35777a894855"},{url:"/_next/static/chunks/3ec2e808-73d3bf460e0bdded.js",revision:"73d3bf460e0bdded"},{url:"/_next/static/chunks/417-fdd1af9de3cc32d6.js",revision:"fdd1af9de3cc32d6"},{url:"/_next/static/chunks/481-1c925e9c846bd912.js",revision:"1c925e9c846bd912"},{url:"/_next/static/chunks/49-ad3409b5a255a2e5.js",revision:"ad3409b5a255a2e5"},{url:"/_next/static/chunks/509-8f8ac6d4279e469e.js",revision:"8f8ac6d4279e469e"},{url:"/_next/static/chunks/518-f8dbcecddaadea02.js",revision:"f8dbcecddaadea02"},{url:"/_next/static/chunks/541-7b98831b4c7e5249.js",revision:"7b98831b4c7e5249"},{url:"/_next/static/chunks/559-fc62b687a83a2ae2.js",revision:"fc62b687a83a2ae2"},{url:"/_next/static/chunks/609-3cb6d5e39bfdc335.js",revision:"3cb6d5e39bfdc335"},{url:"/_next/static/chunks/638-51fbf8ed20a7f717.js",revision:"51fbf8ed20a7f717"},{url:"/_next/static/chunks/670-75c14c0da7064d32.js",revision:"75c14c0da7064d32"},{url:"/_next/static/chunks/736-a0413aae743c0f82.js",revision:"a0413aae743c0f82"},{url:"/_next/static/chunks/737-fe18d559328f1e85.js",revision:"fe18d559328f1e85"},{url:"/_next/static/chunks/80-bb572c3a73622d8c.js",revision:"bb572c3a73622d8c"},{url:"/_next/static/chunks/926-82c4df181b8f380b.js",revision:"82c4df181b8f380b"},{url:"/_next/static/chunks/975-66848ce95fefd414.js",revision:"66848ce95fefd414"},{url:"/_next/static/chunks/977-4a255126c91a75b3.js",revision:"4a255126c91a75b3"},{url:"/_next/static/chunks/app/(authed)/dashboard/analytics/page-7995d9311b7ee4a5.js",revision:"7995d9311b7ee4a5"},{url:"/_next/static/chunks/app/(authed)/dashboard/export-data/page-341a437ec3b6cd2f.js",revision:"341a437ec3b6cd2f"},{url:"/_next/static/chunks/app/(authed)/dashboard/goals/page-d02a3f48809b4532.js",revision:"d02a3f48809b4532"},{url:"/_next/static/chunks/app/(authed)/dashboard/layout-c504d1cdb43b7bd8.js",revision:"c504d1cdb43b7bd8"},{url:"/_next/static/chunks/app/(authed)/dashboard/overview/loading-b6360270ed634ac9.js",revision:"b6360270ed634ac9"},{url:"/_next/static/chunks/app/(authed)/dashboard/overview/page-97ff4e86232b1722.js",revision:"97ff4e86232b1722"},{url:"/_next/static/chunks/app/(authed)/dashboard/page-c3377f4781546314.js",revision:"c3377f4781546314"},{url:"/_next/static/chunks/app/(authed)/dashboard/sessions/page-56317a80a8dc48d1.js",revision:"56317a80a8dc48d1"},{url:"/_next/static/chunks/app/(authed)/layout-975dc9d6cbd95991.js",revision:"975dc9d6cbd95991"},{url:"/_next/static/chunks/app/(public)/login/page-8fc0e6229fb3cc47.js",revision:"8fc0e6229fb3cc47"},{url:"/_next/static/chunks/app/(public)/signup/page-50e6ab48ba6a18e6.js",revision:"50e6ab48ba6a18e6"},{url:"/_next/static/chunks/app/(public)/verify-email/page-afd257b48533005f.js",revision:"afd257b48533005f"},{url:"/_next/static/chunks/app/_not-found/page-093fdcd534118dc9.js",revision:"093fdcd534118dc9"},{url:"/_next/static/chunks/app/layout-af2e34b61aade951.js",revision:"af2e34b61aade951"},{url:"/_next/static/chunks/app/manifest.webmanifest/route-c3377f4781546314.js",revision:"c3377f4781546314"},{url:"/_next/static/chunks/app/page-a865c7ec608b207a.js",revision:"a865c7ec608b207a"},{url:"/_next/static/chunks/f80b594b-4a61aa20d3556417.js",revision:"4a61aa20d3556417"},{url:"/_next/static/chunks/framework-a7169b6999657e0a.js",revision:"a7169b6999657e0a"},{url:"/_next/static/chunks/main-806699314570b1bb.js",revision:"806699314570b1bb"},{url:"/_next/static/chunks/main-app-8c218c77ddd3da1b.js",revision:"8c218c77ddd3da1b"},{url:"/_next/static/chunks/pages/_app-f5c88713886d4d61.js",revision:"f5c88713886d4d61"},{url:"/_next/static/chunks/pages/_error-3acc7e441e81c8b5.js",revision:"3acc7e441e81c8b5"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-6c1008301e9f5c58.js",revision:"6c1008301e9f5c58"},{url:"/_next/static/css/70bf159fd5f9a0a4.css",revision:"70bf159fd5f9a0a4"},{url:"/_next/static/css/de70bee13400563f.css",revision:"de70bee13400563f"},{url:"/_next/static/media/4cf2300e9c8272f7-s.p.woff2",revision:"18bae71b1e1b2bb25321090a3b563103"},{url:"/_next/static/media/747892c23ea88013-s.woff2",revision:"a0761690ccf4441ace5cec893b82d4ab"},{url:"/_next/static/media/8d697b304b401681-s.woff2",revision:"cc728f6c0adb04da0dfcb0fc436a8ae5"},{url:"/_next/static/media/93f479601ee12b01-s.p.woff2",revision:"da83d5f06d825c5ae65b7cca706cb312"},{url:"/_next/static/media/9610d9e46709d722-s.woff2",revision:"7b7c0ef93df188a852344fc272fc096b"},{url:"/_next/static/media/ba015fad6dcf6784-s.woff2",revision:"8ea4f719af3312a055caf09f34c89a77"},{url:"/file.svg",revision:"d09f95206c3fa0bb9bd9fefabfd0ea71"},{url:"/globe.svg",revision:"2aaafa6a49b6563925fe440891e32717"},{url:"/icon-192x192.png",revision:"9d9216467d3d495ba24c45a12c13c4c8"},{url:"/icon-512x512.png",revision:"3ef7bf56a73ac0c01df2ef402e7195ee"},{url:"/next.svg",revision:"8e061864f388b47f33a1c3780831193e"},{url:"/vercel.svg",revision:"c0af2f507b369b085b35ef4bbe3bcf1e"},{url:"/window.svg",revision:"a2760511c65806022ad20adf74370ff3"}],{ignoreURLParametersMatching:[/^utm_/,/^fbclid$/]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:function(e){var s=e.response;return _async_to_generator(function(){return _ts_generator(this,function(e){return[2,s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s]})})()}}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:2592e3})]}),"GET"),e.registerRoute(/\/_next\/static.+\.js$/i,new e.CacheFirst({cacheName:"next-static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4|webm)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:48,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(function(e){var s=e.sameOrigin,a=e.url.pathname;return!(!s||a.startsWith("/api/auth/callback")||!a.startsWith("/api/"))},new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(function(e){var s=e.request,a=e.url.pathname,c=e.sameOrigin;return"1"===s.headers.get("RSC")&&"1"===s.headers.get("Next-Router-Prefetch")&&c&&!a.startsWith("/api/")},new e.NetworkFirst({cacheName:"pages-rsc-prefetch",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(function(e){var s=e.request,a=e.url.pathname,c=e.sameOrigin;return"1"===s.headers.get("RSC")&&c&&!a.startsWith("/api/")},new e.NetworkFirst({cacheName:"pages-rsc",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(function(e){var s=e.url.pathname;return e.sameOrigin&&!s.startsWith("/api/")},new e.NetworkFirst({cacheName:"pages",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(function(e){return!e.sameOrigin},new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")});
+if(!self.define){let e,s={};const a=(a,c)=>(a=new URL(a+".js",c).href,s[a]||new Promise(s=>{if("document"in self){const e=document.createElement("script");e.src=a,e.onload=s,document.head.appendChild(e)}else e=a,importScripts(a),s()}).then(()=>{let e=s[a];if(!e)throw new Error(`Module ${a} didn’t register its module`);return e}));self.define=(c,n)=>{const t=e||("document"in self?document.currentScript.src:"")||location.href;if(s[t])return;let i={};const r=e=>a(e,t),f={module:{uri:t},exports:i,require:r};s[t]=Promise.all(c.map(e=>f[e]||r(e))).then(e=>(n(...e),i))}}define(["./workbox-3c9d0171"],function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/YDv3GfZm4LzwRjOdnmzvi/_buildManifest.js",revision:"e4f1901bff6185f2a0bd4f7b58c33fa1"},{url:"/_next/static/YDv3GfZm4LzwRjOdnmzvi/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/1094-f703795b1be57669.js",revision:"f703795b1be57669"},{url:"/_next/static/chunks/1417-6b87200bc6cf5cbf.js",revision:"6b87200bc6cf5cbf"},{url:"/_next/static/chunks/161cdf59-d3efd7f61c6e213c.js",revision:"d3efd7f61c6e213c"},{url:"/_next/static/chunks/3080-60510f8ce5435fa8.js",revision:"60510f8ce5435fa8"},{url:"/_next/static/chunks/3282-5f2947012ec73956.js",revision:"5f2947012ec73956"},{url:"/_next/static/chunks/3286-6973267246ac81a2.js",revision:"6973267246ac81a2"},{url:"/_next/static/chunks/3559-4c6840595167d553.js",revision:"4c6840595167d553"},{url:"/_next/static/chunks/3771-733966e502e15ece.js",revision:"733966e502e15ece"},{url:"/_next/static/chunks/3798-7b79ba79feb74a3c.js",revision:"7b79ba79feb74a3c"},{url:"/_next/static/chunks/3975-82283f716822db4f.js",revision:"82283f716822db4f"},{url:"/_next/static/chunks/3ec2e808-73d3bf460e0bdded.js",revision:"73d3bf460e0bdded"},{url:"/_next/static/chunks/4467-522f0f7605dd0bcf.js",revision:"522f0f7605dd0bcf"},{url:"/_next/static/chunks/4518-5ea537630f6f0ee0.js",revision:"5ea537630f6f0ee0"},{url:"/_next/static/chunks/4888-cb2097f4839f8240.js",revision:"cb2097f4839f8240"},{url:"/_next/static/chunks/4979-7b2af4b7a604cd36.js",revision:"7b2af4b7a604cd36"},{url:"/_next/static/chunks/5443-74ccfe1156c1b3da.js",revision:"74ccfe1156c1b3da"},{url:"/_next/static/chunks/5499-d7513da2abe1e79b.js",revision:"d7513da2abe1e79b"},{url:"/_next/static/chunks/5796-227be540e76f7b91.js",revision:"227be540e76f7b91"},{url:"/_next/static/chunks/6217-31deda969dbf5ffd.js",revision:"31deda969dbf5ffd"},{url:"/_next/static/chunks/6481-36fb104b7971839d.js",revision:"36fb104b7971839d"},{url:"/_next/static/chunks/6541-6a3f880e51af14f7.js",revision:"6a3f880e51af14f7"},{url:"/_next/static/chunks/6638-2b164f5b3394f2be.js",revision:"2b164f5b3394f2be"},{url:"/_next/static/chunks/8483-56ad6c4b25c2fc18.js",revision:"56ad6c4b25c2fc18"},{url:"/_next/static/chunks/8613-092d1aed59209708.js",revision:"092d1aed59209708"},{url:"/_next/static/chunks/9371-bfc439a04b8de368.js",revision:"bfc439a04b8de368"},{url:"/_next/static/chunks/977-4a255126c91a75b3.js",revision:"4a255126c91a75b3"},{url:"/_next/static/chunks/app/(authed)/dashboard/analytics/page-47a51f0360ed1eb4.js",revision:"47a51f0360ed1eb4"},{url:"/_next/static/chunks/app/(authed)/dashboard/export-data/page-005dae659d888a3f.js",revision:"005dae659d888a3f"},{url:"/_next/static/chunks/app/(authed)/dashboard/goals/page-c769c10679d6da6d.js",revision:"c769c10679d6da6d"},{url:"/_next/static/chunks/app/(authed)/dashboard/layout-d78a0e19d2414988.js",revision:"d78a0e19d2414988"},{url:"/_next/static/chunks/app/(authed)/dashboard/overview/loading-bcf07da1632a1d74.js",revision:"bcf07da1632a1d74"},{url:"/_next/static/chunks/app/(authed)/dashboard/overview/page-e4761cd807dd9e3a.js",revision:"e4761cd807dd9e3a"},{url:"/_next/static/chunks/app/(authed)/dashboard/page-44be739dc6a9acfa.js",revision:"44be739dc6a9acfa"},{url:"/_next/static/chunks/app/(authed)/dashboard/sessions/page-84a31c2a45863e96.js",revision:"84a31c2a45863e96"},{url:"/_next/static/chunks/app/(authed)/layout-d0eee55cd40d43ef.js",revision:"d0eee55cd40d43ef"},{url:"/_next/static/chunks/app/(public)/login/page-7c6bbd3e21558b6d.js",revision:"7c6bbd3e21558b6d"},{url:"/_next/static/chunks/app/(public)/signup/page-21e08a8fb659bed1.js",revision:"21e08a8fb659bed1"},{url:"/_next/static/chunks/app/(public)/verify-email/page-83729559a403a445.js",revision:"83729559a403a445"},{url:"/_next/static/chunks/app/_not-found/page-b7da4bafb372461b.js",revision:"b7da4bafb372461b"},{url:"/_next/static/chunks/app/layout-63ce26ef6f83021a.js",revision:"63ce26ef6f83021a"},{url:"/_next/static/chunks/app/manifest.webmanifest/route-44be739dc6a9acfa.js",revision:"44be739dc6a9acfa"},{url:"/_next/static/chunks/app/manifesto/page-df7405a54c42c6a5.js",revision:"df7405a54c42c6a5"},{url:"/_next/static/chunks/app/page-df7405a54c42c6a5.js",revision:"df7405a54c42c6a5"},{url:"/_next/static/chunks/f80b594b-84bdf96808472cf4.js",revision:"84bdf96808472cf4"},{url:"/_next/static/chunks/framework-d2068cc3fada2cf5.js",revision:"d2068cc3fada2cf5"},{url:"/_next/static/chunks/main-42b25215a5ba0ae5.js",revision:"42b25215a5ba0ae5"},{url:"/_next/static/chunks/main-app-52c677c9b964e527.js",revision:"52c677c9b964e527"},{url:"/_next/static/chunks/pages/_app-2a92b90f93d2ea4a.js",revision:"2a92b90f93d2ea4a"},{url:"/_next/static/chunks/pages/_error-c575b73f1376c123.js",revision:"c575b73f1376c123"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-1eeb70289e73182d.js",revision:"1eeb70289e73182d"},{url:"/_next/static/css/1e50034ab13bc5f8.css",revision:"1e50034ab13bc5f8"},{url:"/_next/static/css/de70bee13400563f.css",revision:"de70bee13400563f"},{url:"/_next/static/media/4cf2300e9c8272f7-s.p.woff2",revision:"18bae71b1e1b2bb25321090a3b563103"},{url:"/_next/static/media/747892c23ea88013-s.woff2",revision:"a0761690ccf4441ace5cec893b82d4ab"},{url:"/_next/static/media/8d697b304b401681-s.woff2",revision:"cc728f6c0adb04da0dfcb0fc436a8ae5"},{url:"/_next/static/media/93f479601ee12b01-s.p.woff2",revision:"da83d5f06d825c5ae65b7cca706cb312"},{url:"/_next/static/media/9610d9e46709d722-s.woff2",revision:"7b7c0ef93df188a852344fc272fc096b"},{url:"/_next/static/media/ba015fad6dcf6784-s.woff2",revision:"8ea4f719af3312a055caf09f34c89a77"},{url:"/file.svg",revision:"d09f95206c3fa0bb9bd9fefabfd0ea71"},{url:"/globe.svg",revision:"2aaafa6a49b6563925fe440891e32717"},{url:"/icon-192x192.png",revision:"9d9216467d3d495ba24c45a12c13c4c8"},{url:"/icon-512x512.png",revision:"3ef7bf56a73ac0c01df2ef402e7195ee"},{url:"/next.svg",revision:"8e061864f388b47f33a1c3780831193e"},{url:"/vercel.svg",revision:"c0af2f507b369b085b35ef4bbe3bcf1e"},{url:"/window.svg",revision:"a2760511c65806022ad20adf74370ff3"}],{ignoreURLParametersMatching:[/^utm_/,/^fbclid$/]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:function(e){var s=e.response;return _async_to_generator(function(){return _ts_generator(this,function(e){return[2,s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s]})})()}}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:2592e3})]}),"GET"),e.registerRoute(/\/_next\/static.+\.js$/i,new e.CacheFirst({cacheName:"next-static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4|webm)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:48,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(function(e){var s=e.sameOrigin,a=e.url.pathname;return!(!s||a.startsWith("/api/auth/callback")||!a.startsWith("/api/"))},new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(function(e){var s=e.request,a=e.url.pathname,c=e.sameOrigin;return"1"===s.headers.get("RSC")&&"1"===s.headers.get("Next-Router-Prefetch")&&c&&!a.startsWith("/api/")},new e.NetworkFirst({cacheName:"pages-rsc-prefetch",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(function(e){var s=e.request,a=e.url.pathname,c=e.sameOrigin;return"1"===s.headers.get("RSC")&&c&&!a.startsWith("/api/")},new e.NetworkFirst({cacheName:"pages-rsc",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(function(e){var s=e.url.pathname;return e.sameOrigin&&!s.startsWith("/api/")},new e.NetworkFirst({cacheName:"pages",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(function(e){return!e.sameOrigin},new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")});
